move ErrorScreen to #/components dir
